backend/requirements.txt
aniso8601==6.0.0
APScheduler==3.6.0
bcrypt==3.1.6
Click==7.0
dnspython==1.16.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.7
gunicorn==19.9.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.1
passlib==1.7.1
pycparser==2.19
pymongo==3.8.0
pyquery==1.4.0
python-dateutil==2.7.3
pytz==2018.7
requests==2.21.0
virtualenv==16.4.1
PyMySQL==0.9.3
mysql-connector==2.2.9
mysqlclient==1.3.12
Flask-MySQL==1.4.0
Flask-MySQLdb==0.2.0
